minor cleanup of localmath function and quittimer behavior

+++ localmath.py
@@ -1,6 +1,6 @@
 def NDCToSC(x, y, screenwidth, screenheight):
     assert x >= 0.0 and x <= 1.0 and y >= 0.0 and y <= 1.0, "Invalid NDCs"
-    return x * screenwidth, y * screenheight
+    return int(x * screenwidth), int(y * screenheight)
 
 def NDCToSC_x(x, screenwidth):
     assert abs(x) >= 0.0 and abs(x) <= 1.0, "Invalid NDCs"
